trip_verbalizer/requirements.txt
# Trip Verbalizer Dependencies
# ============================

# Core async HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Geospatial libraries
geopy>=2.4.0
shapely>=2.0.0

# Numerical computing
numpy>=1.26.0

# Configuration
pyyaml>=6.0.1

# Caching
aiosqlite>=0.19.0
diskcache>=5.6.0

# Rate limiting
aiolimiter>=1.1.0

# CLI and utilities
click>=8.1.0
rich>=13.7.0

# Logging
structlog>=23.2.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0
